requirements.txt
discord.py
flask
requests
playwright
beautifulsoup4